@@ -1,55 +1,64 @@
 {
   "name": "portfolio",
-  "version": "0.1.0",
+  "version": "2.0.0",
+  "author": "Marcos Lopez",
+  "description": "Marcos Lopez Portfolio website",
+  "private": true,
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "serve:ssr:testing": "node dist/testing/server/server.mjs"
+    "serve:ssr:testing": "node dist/portfolio/server/server.mjs",
+    "lint": "ng lint"
   },
-  "private": true,
   "dependencies": {
-    "@angular/common": "18.0.3",
-    "@angular/compiler": "18.0.3",
-    "@angular/core": "18.0.3",
-    "@angular/forms": "18.0.3",
-    "@angular/platform-browser": "18.0.3",
-    "@angular/platform-browser-dynamic": "18.0.3",
-    "@angular/platform-server": "^18.0.3",
-    "@angular/router": "18.0.3",
-    "@angular/ssr": "^18.0.4",
+    "@angular/animations": "18.2.0",
+    "@angular/common": "18.2.0",
+    "@angular/compiler": "18.2.0",
+    "@angular/core": "18.2.0",
+    "@angular/forms": "18.2.0",
+    "@angular/platform-browser": "18.2.0",
+    "@angular/platform-browser-dynamic": "18.2.0",
+    "@angular/platform-server": "18.2.0",
+    "@angular/router": "18.2.0",
+    "@angular/ssr": "18.2.0",
     "aos": "2.3.4",
-    "atropos": "^2.0.2",
-    "express": "^4.19.2",
+    "atropos": "2.0.2",
+    "express": "4.19.2",
     "ng-particles": "3.12.0",
     "rxjs": "7.8.1",
     "tslib": "2.6.3",
     "tsparticles-engine": "2.12.0",
     "tsparticles-slim": "2.12.0",
     "vanilla-tilt": "1.8.1",
-    "zone.js": "0.14.7"
+    "zone.js": "0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "18.0.4",
-    "@angular/cli": "18.0.4",
-    "@angular/compiler-cli": "18.0.3",
+    "@angular-devkit/build-angular": "18.2.0",
+    "@angular/cli": "18.2.0",
+    "@angular/compiler-cli": "18.2.0",
     "@types/aos": "3.0.7",
-    "@types/express": "^4.17.21",
+    "@types/express": "4.17.21",
     "@types/jasmine": "5.1.4",
-    "@types/node": "^20.14.2",
-    "autoprefixer": "10.4.19",
+    "@types/node": "20.16.1",
+    "angular-eslint": "18.3.0",
+    "autoprefixer": "10.4.20",
+    "eslint": "^9.9.0",
     "jasmine-core": "5.1.2",
     "karma": "6.4.3",
     "karma-chrome-launcher": "3.2.0",
     "karma-coverage": "2.2.1",
     "karma-jasmine": "5.1.0",
     "karma-jasmine-html-reporter": "2.1.0",
-    "postcss": "8.4.38",
-    "tailwindcss": "3.4.4",
-    "typescript": "5.4.5"
+    "postcss": "8.4.41",
+    "tailwindcss": "3.4.10",
+    "typescript": "5.4.5",
+    "typescript-eslint": "8.1.0"
   },
-  "keywords": [],
-  "author": "Marcos Lopez"
+  "engines": {
+    "node": ">=20.7.0"
+  },
+  "keywords": []
 }
