@@ -1,14 +1,7 @@
 import { CommonModule } from '@angular/common';
-import {
-  ChangeDetectionStrategy,
-  Component,
-  OnInit,
-  inject,
-} from '@angular/core';
-import { Title } from '@angular/platform-browser';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { AboutIntroLayoutComponent } from '@app/components/legacy/introduction/introduction.component';
-import { StudiesLayoutComponent } from '@app/pages/home/layouts/studies/studies.component';
-import { SeoService } from '@app/services/legacy/seo.service';
+import { StudiesLayout } from '@app/pages/home/layouts/studies/studies.component';
 
 @Component({
   standalone: true,
@@ -16,20 +9,6 @@ import { SeoService } from '@app/services/legacy/seo.service';
   templateUrl: './about.component.html',
   styleUrls: ['./about.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [CommonModule, AboutIntroLayoutComponent, StudiesLayoutComponent],
+  imports: [CommonModule, AboutIntroLayoutComponent, StudiesLayout],
 })
-export class AboutPageComponent implements OnInit {
-  private seo: SeoService = inject(SeoService);
-  private title: Title = inject(Title);
-  constructor() {}
-
-  ngOnInit(): void {
-    let t: string = 'About me - Marcos Lopez Web Portfolio';
-    this.title.setTitle(t);
-    this.seo.generateTags({
-      title: t,
-      description: 'About me page in Marcos López web portfolio',
-      slug: 'about',
-    });
-  }
-}
+export class AboutPageComponent {}
