@@ -1,7 +1,4 @@
-/* tslint:disable:no-unused-variable */
 import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
 
 import { FooterContactComponent } from './footer-contact.component';
 
@@ -11,9 +8,8 @@ describe('FooterContactComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ FooterContactComponent ]
-    })
-    .compileComponents();
+      declarations: [FooterContactComponent],
+    }).compileComponents();
   }));
 
   beforeEach(() => {
