@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
   standalone: true,
@@ -8,10 +8,5 @@ import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
   styleUrls: ['./hero-img.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class HeroImgComponent implements OnInit {
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
+export class HeroImgComponent {
 }
