@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { ServiceCardComponent } from '@app/components/legacy/service-card/service-card.component';
 import { TitleComponent } from '@app/components/atoms/title/title.component';
 import { SERVICES } from 'src/constants/appConst';
@@ -12,14 +12,12 @@ import { SERVICES } from 'src/constants/appConst';
   styleUrls: ['./ourService.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ServiceLayoutComponent implements OnInit {
-  protected readonly SERVICES: any = SERVICES;
+export class ServiceLayoutComponent {
+  protected readonly SERVICES = SERVICES;
   protected readonly titleID: string = 'df4f9aa1';
-  constructor() {}
 
-  ngOnInit(): void {}
 
-  protected trackByFn(index: number, data: any): number | string {
+  protected trackByFn(_: number, data: { ID: string }): number | string {
     return data.ID;
   }
 }
