@@ -1,7 +1,4 @@
-/* tslint:disable:no-unused-variable */
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
+import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProjectCardBodyComponent } from './project-card-body.component';
 
@@ -9,7 +6,7 @@ describe('ProjectCardBodyComponent', () => {
   let component: ProjectCardBodyComponent;
   let fixture: ComponentFixture<ProjectCardBodyComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ProjectCardBodyComponent ]
     })
