@@ -4,20 +4,20 @@ import { ServiceCardComponent } from "./service-card.component";
 describe('ServiceCardComponent', () => {
     let component: ServiceCardComponent;
     let fixture: ComponentFixture<ServiceCardComponent>;
-  
+
     beforeEach(waitForAsync(() => {
       TestBed.configureTestingModule({
         declarations: [ ServiceCardComponent ]
       })
       .compileComponents();
     }));
-  
+
     beforeEach(() => {
       fixture = TestBed.createComponent(ServiceCardComponent);
       component = fixture.componentInstance;
       fixture.detectChanges();
     });
-  
+
     it('should create', () => {
       expect(component).toBeTruthy();
     });
