@@ -1,5 +1,5 @@
 import { NgFor } from '@angular/common';
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { BentoItemComponent } from '@atoms/bento-item/bento-item.component';
 import { ABOUT_USER } from '@constants/userConst';
 import { ExpirenceItemComponent } from '@molecules/expirence-item/expirence-item.component';
