@@ -1,9 +1,4 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  input,
-  Input,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { ProjectArticleImgComponent } from '@app/components/molecules/project-article-img/project-article-img.component';
 import { Project } from '@app/models/projects';
 
