@@ -1,3 +1,3 @@
 import { Type } from '@angular/core';
 
-export type MapIconComponents = { [key: string]: Type<any> };
+export type MapIconComponents = Record<string, Type<unknown>>;
