@@ -5,14 +5,13 @@ import {
   DestroyRef,
   inject,
 } from '@angular/core';
-import { Observable, catchError, firstValueFrom } from 'rxjs';
+import { Observable, catchError } from 'rxjs';
 import { ContactState } from '@app/models/contactState.model';
 import { ToastService } from '@app/lib/toast/Toast.service';
 import { CommonModule } from '@angular/common';
 import { TitleComponent } from '@app/components/atoms/title/title.component';
 import { LiquidBannerComponent } from '@app/components/atoms/liquid-banner/liquid-banner.component';
 import { ContactFormComponent } from '@app/components/organism/contact-form/contact-form.component';
-import { ToastComponent } from '@app/components/organism/toast/toast.component';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 @Component({
@@ -28,7 +27,7 @@ import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
     ContactFormComponent,
   ],
 })
-export class ContactLayoutComponent {
+export class ContactLayout {
   private readonly toast = inject(ToastService);
   private http = inject(HttpClient);
   private destroy = inject(DestroyRef);
@@ -37,7 +36,7 @@ export class ContactLayoutComponent {
   protected handleFormSubmit(event: ContactState) {
     this.toast.info('Sending...', 'We are processing your request');
 
-    let contactForm = new FormData();
+    const contactForm = new FormData();
     contactForm.append('name', event.name);
     contactForm.append('email', event.email);
     contactForm.append('message', event.message);
@@ -45,7 +44,7 @@ export class ContactLayoutComponent {
     this.sendForm(contactForm)
       .pipe(
         takeUntilDestroyed(this.destroy),
-        catchError((error: any) => {
+        catchError((error) => {
           this.toast.error(
             'Error',
             'An unexpected error has occurred with the server'
@@ -54,7 +53,7 @@ export class ContactLayoutComponent {
           return error;
         })
       )
-      .subscribe((response: any) => {
+      .subscribe((_) => {
         this.toast.success(
           'Success',
           'Your message has been sent successfully'
@@ -62,7 +61,7 @@ export class ContactLayoutComponent {
       });
   }
 
-  private sendForm(data: FormData): Observable<any> {
+  private sendForm(data: FormData): Observable<unknown> {
     return this.http.post('https://formspree.io/f/xdorrewr', data, {
       headers: {
         Accept: 'application/json',
