@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { HeroSocialBarComponent } from '@organism/hero-social-bar/hero-social-bar.component';
 import { ArrowDownComponent } from '@atoms/arrow-down/arrow-down.component';
 import { USER } from '@constants/userConst';
@@ -11,6 +11,6 @@ import { USER } from '@constants/userConst';
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [HeroSocialBarComponent, ArrowDownComponent],
 })
-export class HeroLayoutComponent {
+export class HeroLayout {
   protected readonly USER = USER;
 }
